+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # node-sketch-templater Changelog
 
 ## Unreleased
+- Added sketch template for new senseBox board
+- Updated the building process to test the new sketch
+- Update travis configuration: Install senseBox core and use Arduino IDE 1.8.5
 
 ## v1.0.8
 - Use Ethernet2.h library in ethernet templates homeEthernet and homeEthernetFeinstaub